python/etl/requirements.txt
# Core dependencies with exact versions for reproducibility
# Update with: pip freeze > requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.0
pandas==2.1.4
pyarrow==14.0.1
python-dotenv==1.0.0

# Note: For production, generate a complete lock file with:
# pip freeze > requirements.lock
# This ensures all transitive dependencies are pinned
